fix: Fix build failure due to ambiguous overloaded function call
This patch fixes the following build error:

```
libtool: compile:  g++ -DHAVE_CONFIG_H -I. -I.. -I../include -O2 -pipe -march=native -MT libimage_stream_la-fax-encoder.lo -MD -MP -MF .deps/libimage_stream_la-fax-encoder.Tpo -c fax-encoder.cc  -fPIC -DPIC -o .libs/libimage_stream_la-fax-encoder.o
jpegstream.cc: In member function ‘virtual iscan::basic_imgstream& iscan::jpegstream::write(const byte_type*, iscan::jpegstream::size_type)’:
jpegstream.cc:85:31: error: call of overloaded ‘div(unsigned int&, long unsigned int)’ is ambiguous
   85 |             div_t index = div (i, 8 * sizeof (JSAMPLE));
      |                           ~~~~^~~~~~~~~~~~~~~~~~~~~~~~~
In file included from /usr/include/c++/11/cstdlib:75,
                 from /usr/include/c++/11/stdlib.h:36,
                 from /usr/include/libltdl/lt_system.h:35,
                 from /usr/include/ltdl.h:36,
                 from basic-imgstream.hh:55,
                 from jpegstream.hh:41,
                 from jpegstream.cc:34:
```

Source: aur.git - AUR Package Repositories <https://aur.archlinux.org/cgit/aur.git/commit/jpegstream.cc.patch?h=iscan&id=32ff91a61c0d1eb7a0a3a757cfaaa2034dd48dc1>
Authored-by: Ivan Zenin

+++ lib/jpegstream.cc
@@ -33,7 +33,7 @@
 
 #include "jpegstream.hh"
 
-#include <cstdlib>
+// #include <cstdlib>
 #include <ios>
 
 namespace iscan
@@ -82,7 +82,8 @@ namespace iscan
         //        only that _bits != 8.
         for (unsigned int i = 0; i < _h_sz; ++i)
           {
-            div_t index = div (i, 8 * sizeof (JSAMPLE));
+            div_t index = div (static_cast<int>(i),
+                               static_cast<int>(8 * sizeof (JSAMPLE)));
             int offset = 8 * sizeof (JSAMPLE) - 1 - index.rem;
             _scanline[i] = ((line[index.quot] & (1 << offset))
                             ? 0 : ~0);
